C++ modernize: replace a loop for an algorithm.
Although this should be less efficient as it is not really
necessary for the problem to materialize the intersection,
only its size is relevant.

I think that an improvement in this direction would be
using ranges views, with a filter and then computing the
size of the resulting range or range view(?).