Day 10! ~7h of grinding work. Mess of a code.
And still even not fully automatic. I am "painting" one side of
the loop one color (x) and another, another one (o). The code
does not detect which part is inside and which is outside. Also,
the painting still does not handle all possible corner cases. That
I know of, in my input I have a . cell surrounded by u, d, ...
that is never painted:

                            ldulu
                            ul.ru
                            rrdud