Playing step by step.
- Use formula for mod of negative numbers iso loop.
- set -> unordered_set with a bit of boost for minimal handcode.
Neither had a sensible impact running with sample input for 500 steps.
It takes several seconds up to a say a minute.
With 1000 steps is not finishing while I commit this.
We may expand the ones that are surely going to continue going far with
math and then do the simulation for the others? Hmm.